requirements.txt
numpy
pandas
streamlit
re
nltk
nltk.corpus
nltk.stem.porter
sklearn.feature_extraction.text
